Relocate `labels` module inside of `github`
